# Conflicts:
#	config/locales/en.yml

+++ app/mutations/task/decline.rb
@@ -7,9 +7,12 @@ class Task::Decline < Mutations::Command
   def execute
     assignment = Task::Role.send('task.volunteer').find_by(task: task, user: user)
 
+    return unless assignment.present?
+
     assignment.destroy
 
     (task.users.uniq - [ user ]).each do |outboud_user|
+      next unless outboud_user.email.present?
       TaskMailer.task_change(task, outboud_user).deliver_now
     end
